Update en.json
Hi,

Not listing everything here, as some are self explanatory, but have included some where I just wanted to provide a bit of context to my proposed change. A lot of the changes are either minor corrections, consistency changes (to keep the wording same or similar throughout the game), or to further clarify what the card does.

I have used gwent.one as reference, as it contains a database of all beta Gwent cards as they appeared upon the last update, pulled directly from the game. I have of course taken the custom balance changes etc. into account, to keep their new functionality / stats.

This should cover the Neutral cards (Down to Dagon at 21001 / Line 880)

I hope at least some of this is useful. and want to thank you for bringing this project to life and giving me a healthy dose of nostalgia!

12004 Geralt of Rivia: Just to keep the word for the card effects / abilities the same throughout the game, for consistency. Old Gwent used ability, and can see it being used on some custom balance changes and cards as well.

12008 Villentretenmerth: Not sure if coded differently from how it originally worked, but if functioning the same as original Gwent, I added a bit of text to clarify when it happens and that it excludes himself (in case he is the highest or tied for). Please feel free to change to the below if you would want this to be worded exactly as old Gwent, I just think my edit reads a bit clearer. (Old Gwent: After 3 turns, destroy all the other Highest units on turn start. 3 Armor.)

12011 Dandelion: Vainglory: Original included Triss, not sure if just a typo, but may be worth checking if included in the code (sorry, I'm not sure how to do that).

12015 Zoltan: Scoundrel: Not sure if you changed the functionality of this one (which as how it was written, seems to be a nerf, unless Zoltan himself had an increase in Power by 4), but if functionality to old Gwent has not changed, then I updated the wording to match Old Gwent.

12018 Ihuarraquax: Fixed wording to indicate how it worked in Old Gwent, otherwise it indicate that it keeps damaging 3 enemy units by 7 at every turn end, when it only does it once. Might be worth checking code if coded to keep firing ability or only once.

12026 Triss: Telekinesis: Changed from current to starting, as that was how it was in Old Gwent. Not sure if coded as such or a custom change?

12027 Aguara: Again, not sure if custom change or not, but is random Elf in Old Gwent.

14007 Dimeritium Shackles: Was this function changed? It implies to only be able to lock cards, when in Old Gwent it could also unlock (toggling).

14013 Mardroeme: Seems two cards were named Spore, and this one was originally called Mardroeme, so fixed it.