Update LonelyChampion.cs
A small hotfix for Lonely Champion, several people complained about it on the Discord and say they might take a break until it is fixed. Unitless decks are always unhealthy for the game and this is beyonf overpowered. I don't want to overstep, this is just a hotfix, you can rebalance the card howerver you want when you do the next update.